requirements.txt
# Azure Functions
azure-functions>=1.17.0

# Bot Framework
botbuilder-core>=4.14.0
botbuilder-schema>=4.14.0

# Microsoft Graph API
azure-identity>=1.15.0
aiohttp>=3.9.0

# Azure Storage (for temp file storage)
azure-storage-blob>=12.19.0

# PDF Processing
PyMuPDF>=1.23.0

# Utilities
python-dotenv>=1.0.0
